+++ src/test/java/uk/gov/pay/api/it/CachingAuthenticatorIT.java
@@ -28,7 +28,7 @@ import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static uk.gov.pay.api.model.TokenPaymentType.CARD;
 import static uk.gov.pay.api.utils.WiremockStubbing.stubPublicAuthV1ApiAuth;
 
-public class CachingAuthenticatorTest {
+public class CachingAuthenticatorIT {
     
     private String accountId = "123";
     private String bearerToken = ApiKeyGenerator.apiKeyValueOf("TEST_BEARER_TOKEN", "qwer9yuhgf");
