+++ src/test/java/uk/gov/pay/api/it/GetDirectDebitEventsIT.java
@@ -23,7 +23,7 @@ import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 
-public class GetDirectDebitEventsTest {
+public class GetDirectDebitEventsIT {
 
     static final String API_KEY = ApiKeyGenerator.apiKeyValueOf("TEST_BEARER_TOKEN", "qwer9yuhgf");
