+++ src/test/java/uk/gov/pay/api/it/validation/PaymentsResourceDescriptionValidationIT.java
@@ -17,7 +17,7 @@ import static javax.ws.rs.core.HttpHeaders.AUTHORIZATION;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.core.Is.is;
 
-public class PaymentsResourceReferenceValidationITest extends PaymentResourceITestBase {
+public class PaymentsResourceDescriptionValidationIT extends PaymentResourceITestBase {
 
     private PublicAuthMockClient publicAuthMockClient = new PublicAuthMockClient(publicAuthMock);
     
@@ -27,12 +27,12 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
     }
 
     @Test
-    public void createPayment_responseWith400_whenReferenceIsNumeric() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsNumeric() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
-                "  \"reference\" : 1234," +
-                "  \"description\" : \"Some description\"," +
+                "  \"description\" : 1234," +
+                "  \"reference\" : \"Some reference\"," +
                 "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
@@ -44,18 +44,18 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("reference"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0102"))
-                .assertThat("$.description", is("Invalid attribute value: reference. Must be a valid string format"));
+                .assertThat("$.description", is("Invalid attribute value: description. Must be a valid string format"));
     }
 
     @Test
-    public void createPayment_responseWith400_whenReferenceIsEmpty() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsEmpty() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
-                "  \"reference\" : \"\"," +
-                "  \"description\" : \"Some description\"," +
+                "  \"description\" : \"\"," +
+                "  \"reference\" : \"Some reference\"," +
                 "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
@@ -67,18 +67,18 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("reference"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0101"))
-                .assertThat("$.description", is("Missing mandatory attribute: reference"));
+                .assertThat("$.description", is("Missing mandatory attribute: description"));
     }
 
     @Test
-    public void createPayment_responseWith400_whenReferenceIsBlank() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsBlank() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
-                "  \"reference\" : \"    \"," +
-                "  \"description\" : \"Some description\"," +
+                "  \"description\" : \"    \"," +
+                "  \"reference\" : \"Some reference\"," +
                 "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
@@ -90,17 +90,17 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("reference"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0101"))
-                .assertThat("$.description", is("Missing mandatory attribute: reference"));
+                .assertThat("$.description", is("Missing mandatory attribute: description"));
     }
 
     @Test
-    public void createPayment_responseWith400_whenReferenceIsMissing() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsMissing() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
-                "  \"description\" : \"Some description\"," +
+                "  \"reference\" : \"Some reference\"," +
                 "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
@@ -112,19 +112,19 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("reference"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0101"))
-                .assertThat("$.description", is("Missing mandatory attribute: reference"));
+                .assertThat("$.description", is("Missing mandatory attribute: description"));
     }
 
     @Test
-    public void createPayment_responseWith400_whenReferenceIsNull() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsNull() throws IOException {
 
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
-                "  \"reference\" : null," +
-                "  \"description\" : \"Some description\"," +
+                "  \"description\" : null," +
+                "  \"reference\" : \"Some reference\"," +
                 "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
@@ -136,20 +136,20 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("reference"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0101"))
-                .assertThat("$.description", is("Missing mandatory attribute: reference"));
+                .assertThat("$.description", is("Missing mandatory attribute: description"));
     }
 
     @Test
-    public void createPayment_responseWith422_whenReferenceSizeIsGreaterThanMaxLength() throws IOException {
+    public void createPayment_responseWith422_whenDescriptionSizeIsGreaterThanMaxLength() throws IOException {
 
         String aVeryLongReference = RandomStringUtils.randomAlphanumeric(256);
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
-                "  \"reference\" : \"" + aVeryLongReference + "\"," +
-                "  \"description\" : \"Some description\"," +
+                "  \"description\" : \"" + aVeryLongReference + "\"," +
+                "  \"reference\" : \"Some reference\"," +
                 "  \"return_url\" : \"https://www.example.com/return_url\"" +
                 "}";
 
@@ -161,18 +161,18 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("reference"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0102"))
-                .assertThat("$.description", is("Invalid attribute value: reference. Must be less than or equal to 255 characters length"));
+                .assertThat("$.description", is("Invalid attribute value: description. Must be less than or equal to 255 characters length"));
     }
 
     @Test
-    public void createPayment_responseWith400_whenReferenceHasNotAValidJsonValue() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionHasNotAValidJsonValue() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
-                "  \"reference\" : " +
-                "  \"description\" : \"Some description\"," +
+                "  \"description\" : " +
+                "  \"reference\" : \"Some reference\"," +
                 "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
@@ -189,12 +189,12 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
     }
 
     @Test
-    public void createPayment_responseWith400_whenReferenceFieldIsNotExpectedJsonField() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionFieldIsNotExpectedJsonField() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
-                "  \"reference\" : {\"whatever\" : 1}," +
-                "  \"description\" : \"Some description\"," +
+                "  \"description\" : {\"whatever\" : 1}," +
+                "  \"reference\" : \"Some reference\"," +
                 "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
@@ -206,9 +206,9 @@ public class PaymentsResourceReferenceValidationITest extends PaymentResourceITe
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("reference"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0102"))
-                .assertThat("$.description", is("Invalid attribute value: reference. Must be a valid string format"));
+                .assertThat("$.description", is("Invalid attribute value: description. Must be a valid string format"));
     }
 
     private ValidatableResponse postPaymentResponse(String bearerToken, String payload) {
